Fixed my bug of state corruption mostly unrelated to concurrency.
I was corrupting my own state (sometimes a vtable) whenever the
division of labor was not evenly distributed because my output table
clearing code was allowing itself to run off the end of the array.
Fixed by clamping the clearing code so that it does not go past the
end.